bytes

+++ src/cryptonote_config.h
@@ -158,16 +158,16 @@ namespace config
   uint64_t const ELECTRONERO_HARDFORK = ((uint64_t)239930); // first planned fork
   uint64_t const ELECTRONERO_HARDFORK_HEIGHT = ((uint64_t)239921); // initial fork height
   
-  uint32_t const MAINNET_HARDFORK_V1_HEIGHT = ((uint32_t)1); // v1
-  uint32_t const MAINNET_HARDFORK_V2_HEIGHT = ((uint32_t)239965); // v2
-  uint32_t const MAINNET_HARDFORK_V3_HEIGHT = ((uint32_t)239970); // v3
-  uint32_t const MAINNET_HARDFORK_V4_HEIGHT = ((uint32_t)239975); // v4
-  uint32_t const MAINNET_HARDFORK_V5_HEIGHT = ((uint32_t)239980); // v5
-  uint32_t const MAINNET_HARDFORK_V6_HEIGHT = ((uint32_t)239930); // v6
-  uint32_t const MAINNET_HARDFORK_V7_HEIGHT = ((uint32_t)239949); // v7
-  uint32_t const MAINNET_HARDFORK_V8_HEIGHT = ((uint32_t)239960); // v8
+  uint64_t const MAINNET_HARDFORK_V1_HEIGHT = ((uint64_t)1); // v1
+  uint64_t const MAINNET_HARDFORK_V2_HEIGHT = ((uint64_t)239965); // v2
+  uint64_t const MAINNET_HARDFORK_V3_HEIGHT = ((uint64_t)239970); // v3
+  uint64_t const MAINNET_HARDFORK_V4_HEIGHT = ((uint64_t)239975); // v4
+  uint64_t const MAINNET_HARDFORK_V5_HEIGHT = ((uint64_t)239980); // v5
+  uint64_t const MAINNET_HARDFORK_V6_HEIGHT = ((uint64_t)239930); // v6
+  uint64_t const MAINNET_HARDFORK_V7_HEIGHT = ((uint64_t)239949); // v7
+  uint64_t const MAINNET_HARDFORK_V8_HEIGHT = ((uint64_t)239960); // v8
   
-  uint8_t const DEFAULT_FEE_ATOMIC_XMR_PER_KB = 5;
+  uint64_t const DEFAULT_FEE_ATOMIC_XMR_PER_KB = 5;
   uint8_t const FEE_CALCULATION_MAX_RETRIES = 10;
   uint64_t const DEFAULT_DUST_THRESHOLD = ((uint64_t)2); // 2 * pow(10, 1)
   uint64_t const BASE_REWARD_CLAMP_THRESHOLD = ((uint64_t)100000000); // pow(10, 8)
